Increase map size to 10, rename quadrant->sector, sector->subsector
Do this so each square is about 2 light years in trek canon assuming each board is now 10x10 squares ("subsectors") - making this change also makes a single subsector a tenth of the board - so that we can move the ship in increments of tenths (.1) instead of eighths (.125) making the mental math nicer - or we can just multiply by 10 and make warp factor input  in terms of subsectors.

Details:
- make string pad take a custom string
- replace hard coded map border code with utility function

TODO adjust font thing again so it all fits on one screen horz and vertically.