Start to refactor torpedo; map footer change
This build not functional.

Original game had targeting computer.  Roll this into  torpedo tube damage function and make it a menu to select the target, with the added bonus of reducing typing.
Torpedo function reacts to damage now, either we can't fire if its too low or if its a little higher the targeting computer functionality is off and we have to do the old school manual targeting.

Make map footer a little more useful by adding the '=' character to match with the header.