Make it so you can click the map grid instead of typing for long range jump
TODO : make the map not clickable (and not have dots) when not in navigation mode

TODO: add a galaxy map screen to click on for long range jumps.  Maybe make this always visible when you do a long range jump even in the typey version?

TODO: make all the textual changes for the mirror vs prime timeline.